# Reduce size of dctx by reutilizing dst buffer