WIP, this round of optimizations has gotten performance much closer to parity, though it has introduced a checksum error in the 270MB file test I'm still tracking down.  This however hasn't affected the smaller size tests; benchmarks indicate that in some cases we now see performance improvements on top of the memory reduction due to the improved cache behavior.  However there's other cases, at low file sizes and high compressibility, where we are still about 1% behind parity.

old performance

./tests/fullbench -b2 -B1000 -P0
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 1000 bytes :                                                  
 2#decompress                   :  4987.5 MB/s  (    1000) 
./tests/fullbench -b2 -B1000 -P10
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 1000 bytes :                                                  
 2#decompress                   :   612.0 MB/s  (    1000) 
./tests/fullbench -b2 -B1000 -P50
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 1000 bytes :                                                  
 2#decompress                   :   585.8 MB/s  (    1000) 
./tests/fullbench -b2 -B1000 -P90
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 1000 bytes :                                                  
 2#decompress                   :  2597.8 MB/s  (    1000) 
./tests/fullbench -b2 -B1000 -P100
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 1000 bytes :                                                  
 2#decompress                   :  2635.7 MB/s  (    1000) 
./tests/fullbench -b2 -B10000 -P0
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 10000 bytes :                                                 
 2#decompress                   : 36167.5 MB/s  (   10000) 
./tests/fullbench -b2 -B10000 -P10
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 10000 bytes :                                                 
 2#decompress                   :  1292.4 MB/s  (   10000) 
./tests/fullbench -b2 -B10000 -P50
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 10000 bytes :                                                 
 2#decompress                   :  1671.9 MB/s  (   10000) 
./tests/fullbench -b2 -B10000 -P90
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 10000 bytes :                                                 
 2#decompress                   :  3205.2 MB/s  (   10000) 
./tests/fullbench -b2 -B10000 -P100
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 10000 bytes :                                                 
 2#decompress                   :  6179.7 MB/s  (   10000) 
./tests/fullbench -b2 -B100000 -P0
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 100000 bytes :                                                
 2#decompress                   : 51880.0 MB/s  (  100000) 
./tests/fullbench -b2 -B100000 -P10
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 100000 bytes :                                                
 2#decompress                   :  1237.1 MB/s  (  100000) 
./tests/fullbench -b2 -B100000 -P50
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 100000 bytes :                                                
 2#decompress                   :  2151.4 MB/s  (  100000) 
./tests/fullbench -b2 -B100000 -P90
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 100000 bytes :                                                
 2#decompress                   :  3193.0 MB/s  (  100000) 
./tests/fullbench -b2 -B100000 -P100
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 100000 bytes :                                                
 2#decompress                   :  7095.5 MB/s  (  100000) 
./tests/fullbench -b2 -B1000000 -P0
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 1000000 bytes :                                               
 2#decompress                   : 34106.0 MB/s  ( 1000000) 
./tests/fullbench -b2 -B1000000 -P10
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 1000000 bytes :                                               
 2#decompress                   :  1309.3 MB/s  ( 1000000) 
./tests/fullbench -b2 -B1000000 -P50
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 1000000 bytes :                                               
 2#decompress                   :  1973.5 MB/s  ( 1000000) 
./tests/fullbench -b2 -B1000000 -P90
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 1000000 bytes :                                               
 2#decompress                   :  2637.8 MB/s  ( 1000000) 
./tests/fullbench -b2 -B1000000 -P100
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 1000000 bytes :                                               
 2#decompress                   : 14852.5 MB/s  ( 1000000) 





new performance

./tests/fullbench -b2 -B1000 -P0
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 1000 bytes :                                                  
 2#decompress                   :  4999.4 MB/s  (    1000) 
./tests/fullbench -b2 -B1000 -P10
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 1000 bytes :                                                  
 2#decompress                   :   609.1 MB/s  (    1000) 
./tests/fullbench -b2 -B1000 -P50
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 1000 bytes :                                                  
 2#decompress                   :   583.5 MB/s  (    1000) 
./tests/fullbench -b2 -B1000 -P90
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 1000 bytes :                                                  
 2#decompress                   :  2402.1 MB/s  (    1000) 
./tests/fullbench -b2 -B1000 -P100
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 1000 bytes :                                                  
 2#decompress                   :  2587.4 MB/s  (    1000) 
./tests/fullbench -b2 -B10000 -P0
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 10000 bytes :                                                 
 2#decompress                   : 37441.8 MB/s  (   10000) 
./tests/fullbench -b2 -B10000 -P10
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 10000 bytes :                                                 
 2#decompress                   :  1297.5 MB/s  (   10000) 
./tests/fullbench -b2 -B10000 -P50
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 10000 bytes :                                                 
 2#decompress                   :  1656.7 MB/s  (   10000) 
./tests/fullbench -b2 -B10000 -P90
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 10000 bytes :                                                 
 2#decompress                   :  3081.0 MB/s  (   10000) 
./tests/fullbench -b2 -B10000 -P100
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 10000 bytes :                                                 
 2#decompress                   :  6127.2 MB/s  (   10000) 
./tests/fullbench -b2 -B100000 -P0
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 100000 bytes :                                                
 2#decompress                   : 52215.9 MB/s  (  100000) 
./tests/fullbench -b2 -B100000 -P10
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 100000 bytes :                                                
 2#decompress                   :  1252.2 MB/s  (  100000) 
./tests/fullbench -b2 -B100000 -P50
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 100000 bytes :                                                
 2#decompress                   :  2146.6 MB/s  (  100000) 
./tests/fullbench -b2 -B100000 -P90
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 100000 bytes :                                                
 2#decompress                   :  3614.6 MB/s  (  100000) 
./tests/fullbench -b2 -B100000 -P100
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 100000 bytes :                                                
 2#decompress                   :  7084.7 MB/s  (  100000) 
./tests/fullbench -b2 -B1000000 -P0
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 1000000 bytes :                                               
 2#decompress                   : 33857.1 MB/s  ( 1000000) 
./tests/fullbench -b2 -B1000000 -P10
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 1000000 bytes :                                               
 2#decompress                   :  1288.9 MB/s  ( 1000000) 
./tests/fullbench -b2 -B1000000 -P50
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 1000000 bytes :                                               
 2#decompress                   :  2095.4 MB/s  ( 1000000) 
./tests/fullbench -b2 -B1000000 -P90
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 1000000 bytes :                                               
 2#decompress                   :  2786.2 MB/s  ( 1000000) 
./tests/fullbench -b2 -B1000000 -P100
*** Zstandard speed analyzer 1.5.0 64-bits, by Yann Collet (Aug 19 2021) ***
 Sample 1000000 bytes :                                               
 2#decompress                   : 15258.4 MB/s  ( 1000000)