@binhdvo 2021-08-24
Still fixing some fuzzer errors that are blocking CI, updated performance numbers below

(All in MB/s)

```
1000 tests:

        orig          new
-P0     5107.3        5264.5 
-P10    628.7         627.9  
-P50    603.3         608.0  
-P90    2665.0        2505.6 
-P100   2688.1        2652.5 

10000 tests:

        orig          new
-P0     37406.7       39007.2
-P10    1328.4        1335.2 
-P50    1715.6        1699.0 
-P90    3259.1        3204.5 
-P100   6408.1        6419.9 

100000 tests:

        orig          new
-P0     52696.0       53962.9
-P10    1263.9        1291.1 
-P50    1801.9        2208.2 
-P90    3255.7        3329.3 
-P100   7338.0        7422.0 

1000000 tests:

        orig          new
-P0     34991.3       34964.2
-P10    1353.2        1337.7 
-P50    2122.0        2158.6 
-P90    2785.3        2777.5 
-P100   15813.2       15834.1
```