@Cyan4973 2021-08-24
I presume you are using the synthetic content generator.
While it's a good way to start investigating this topic,
consider also measuring some "real life" content.

`enwik` is a good example of "real" content that is different from the generator, as it's full of small overlapping matches, which is something the synthetic generator doesn't emulate well.

Also, look at individual files (_not_ the whole package) in `calgary.tar` or even `silesia.tar`, to see if some of them "stand out".