@senhuang42 2021-08-24
The memory reductions are nice!

I took a quick look at this on `enwik7` with `perf`, (using the command, `perf stat -e "cycles,branches,branch-misses,instructions,page-faults,L1-dcache-load-misses,L1-dcache-loads,L1-dcache-stores" -- ./zstd -b1 ../data/enwik7`) and got the following results:

**dev**
```
 1#enwik7            : 9.54 MiB -> 3.92 MiB (2.432),  262.2 MB/s, 1039.5 MB/s 

 Performance counter stats for './zstd -b1 ../data/enwik7':

    22,422,771,549      cycles                                                        (71.43%)
     5,897,438,746      branches                                                      (57.15%)
       159,972,762      branch-misses             #    2.71% of all branches          (57.15%)
    59,148,636,696      instructions              #    2.64  insn per cycle           (71.43%)
             7,523      page-faults                                                 
       925,264,639      L1-dcache-load-misses     #    5.24% of all L1-dcache accesses  (71.43%)
    17,655,116,114      L1-dcache-loads                                               (71.42%)
     7,476,309,120      L1-dcache-stores                                              (71.42%)

       6.237735345 seconds time elapsed

       6.214009000 seconds user
       0.009980000 seconds sys
```
**binhdctx2**
```
 1#enwik7            : 9.54 MiB -> 3.92 MiB (2.432),  262.1 MB/s,  901.9 MB/s 

 Performance counter stats for './zstd -b1 ../data/enwik7':

    22,577,928,991      cycles                                                        (71.44%)
     5,655,449,556      branches                                                      (57.15%)
       152,349,238      branch-misses             #    2.69% of all branches          (57.15%)
    55,533,277,104      instructions              #    2.46  insn per cycle           (71.44%)
             7,508      page-faults                                                 
       915,748,699      L1-dcache-load-misses     #    5.54% of all L1-dcache accesses  (71.44%)
    16,533,161,474      L1-dcache-loads                                               (71.41%)
     6,713,694,511      L1-dcache-stores                                              (71.41%)

       6.280837396 seconds time elapsed

       6.257051000 seconds user
       0.009977000 seconds sys
```
The stats unfortunately are not that helpful, since it seems that instructions per cycle decreases without a very clear culprit, but it seems to indicate a 10% decompression speed regression for me.