@binhdvo 2021-08-31
Some further performance improvements, the new version fairly consistently wins at lower compression levels, and now has some wins and losses at higher compression levels across enwik and silesia.

```
Silesia+enwik corpus tests level 1

            orig     new
enwik7.txt  1116.6   1131.1
dickens     1142.3   1141.8
mozilla     1120.0   1143.5
mr          1347.9   1351.2
nci         2000.1   2020.3
ooffice     927.3    956.0
osdb        1389.3   1398.3
reymont     1156.4   1164.5
samba       1548.0   1583.2
sao         1017.6   1002.6
webster     1178.4   1199.2
xml         1908.8   1949.5
x-ray       1026.4   1039.0


Silesia+enwik corpus tests level 10

            orig     new
enwik7.txt  912.2    919.1
dickens     908.9    867.8
mozilla     1009.2   1044.7
mr          921.9    899.8
nci         2387.6   2411.5
ooffice     783.8    788.9
osdb        1232.5   1173.6
reymont     1105.9   1090.7
samba       1558.1   1556.2
sao         800.9    785.8
webster     1040.8   1027.8
xml         2224.9   2214.4
x-ray       553.5    527.2


Silesia+enwik corpus tests level 22

            orig     new
enwik7.txt  1028.8   1037.5
dickens     1047.3   994.8
mozilla     861.1    846.1
mr          1029.0   969.4
nci         2781.5   2831.2
ooffice     684.0    684.9
osdb        1412.9   1443.7
reymont     1332.0   1278.4
samba       1602.1   1559.6
sao         831.1    837.5
webster     1133.6   1126.7
xml         2513.6   2499.7
x-ray       607.7    599.3
```