@binhdvo 2021-09-07
All fuzzing test issues have been fixed, CI tests now pass.  Performance looks generally good on Mac, with consistent performance improvements on top of the memory savings across enwik7 and silesia at default compression levels.  At maximum compression levels there are some wins and some losses.  On the devserver there were initially some stark losses, tweaks to alignment across the two loops and scratch buffer size have improved these losses but they are still present.

```
Silesia+enwik corpus tests level 1, Mac

            orig     new
enwik7.txt  1085.9   1095.2
dickens     1159.2   1196.7
mozilla     1135.1   1151.6
mr          1371.6   1384.0
nci         2030.5   2049.8
ooffice     950.8    966.7
osdb        1389.1   1429.2
reymont     1096.6   1236.1
samba       1625.5   1670.4
sao         1062.4   1008.4
webster     1245.3   1285.2
xml         2019.7   2059.3
x-ray       1049.0   1049.2


Silesia+enwik corpus tests level 22, Mac

            orig     new
enwik7.txt  962.9    985.2
dickens     1041.6   1000.9
mozilla     875.5    885.1
mr          988.0    963.1
nci         2763.7   2928.6
ooffice     683.8    695.7
osdb        1418.9   1404.4
reymont     1299.4   1281.7
samba       1564.4   1567.3
sao         813.3    815.7
webster     1134.3   1091.5
xml         2544.2   2522.8
x-ray       606.3    611.9


Silesia+enwik corpus tests level 1, devvm (57GB)

            orig     new
enwik7.txt  730.5    705.4
dickens     744.6    719.7
mozilla     715.1    714.3
mr          892.7    872.1
nci         1143.0   1120.1
ooffice     663.0    652.2
osdb        848.2    827.1
reymont     707.0    721.5
samba       972.0    947.4
sao         633.7    599.1
webster     746.2    718.6
xml         1346.2   1344.0
x-ray       739.7    738.0


Silesia+enwik corpus tests level 22, devvm (57GB)

            orig     new
enwik7.txt  538.7    535.6
dickens     414.9    494.1
mozilla     459.2    420.6
mr          441.8    344.6
nci         1291.3   1212.6
ooffice     402.1    385.3
osdb        619.4    466.1
reymont     661.9    709.5
samba       851.0    791.4
sao         421.6    400.3
webster     372.4    347.8
xml         1737.8   1675.8
x-ray       266.5    318.7

```